requirements.txt
python-dotenv
pydantic-ai
pydantic-ai[logfire]
pydantic-settings
logfire
